+++ udata/core/spatial/commands.py
@@ -125,6 +125,12 @@ def load(filename=DEFAULT_GEOZONES_FILE, drop=False):
     shutil.rmtree(tmp.path('translations'))  # Not in use for now.
 
 
+def safe_tarinfo(tarinfo):
+    '''make a tarinfo utf8-compatible'''
+    tarinfo.name = tarinfo.name.decode('utf8')
+    return tarinfo
+
+
 @grp.command()
 @click.argument('filename', metavar='<filename>')
 def load_logos(filename):
@@ -139,8 +145,9 @@ def load_logos(filename):
 
     log.info('Extracting GeoLogos bundle')
     with contextlib.closing(lzma.LZMAFile(filename)) as xz:
-        with tarfile.open(fileobj=xz) as f:
-            f.extractall(tmp.root)
+        with tarfile.open(fileobj=xz, encoding='utf8') as tar:
+            decoded = (safe_tarinfo(t) for t in tar.getmembers())
+            tar.extractall(tmp.root, members=decoded)
 
     log.info('Moving to the final location and cleaning up')
     if os.path.exists(logos.root):