+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Don't automatically overwrite `last_update` field if manually set [#2020](https://github.com/opendatateam/udata/pull/2220)
 - Spatial completion: only index last version of each zone and prevent completion cluttering [#2140](https://github.com/opendatateam/udata/pull/2140)
 - Init: prompt to loads countries [#2140](https://github.com/opendatateam/udata/pull/2140)
+- Handle UTF-8 filenames in `spatial load_logos` command [#2223](https://github.com/opendatateam/udata/pull/2223)
 
 ## 1.6.12 (2019-06-26)